Move Util classes to framework-cloud-common (rebase)
(cherry picked from commit 3e0cf2058e990ea9a030a7b5459188c8d06edf95)
