+++ test-cloud/test-cloud-common/pom.xml
@@ -14,7 +14,17 @@
 
   <name>KIE :: Cloud :: Common Test</name>
   <description>Module providing test utilities and parent test classes.</description>
-
+  
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>org.uberfire</groupId>
+        <artifactId>uberfire-rest-client</artifactId>
+        <version>${version.org.uberfire}</version>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
+      
   <dependencies>
     <dependency>
       <groupId>org.kie.cloud</groupId>
@@ -27,6 +37,50 @@
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.server</groupId>
+      <artifactId>kie-server-client</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.server</groupId>
+      <artifactId>kie-server-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.server</groupId>
+      <artifactId>kie-server-controller-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.server</groupId>
+      <artifactId>kie-server-controller-client</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.server</groupId>
+      <artifactId>kie-server-integ-tests-common</artifactId>
+      <exclusions>
+        <exclusion>
+          <!-- Collides with 'xml-apis:xml-apis' -->
+          <groupId>javax.xml.stream</groupId>
+          <artifactId>stax-api</artifactId>
+        </exclusion>
+      </exclusions>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.server</groupId>
+      <artifactId>kie-server-router-proxy</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie</groupId>
+      <artifactId>business-central-tests-rest</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.cloud</groupId>
+      <artifactId>framework-cloud-common</artifactId>
+    </dependency>
+    <!-- UberFire rest client -->
+    <dependency>
+      <groupId>org.uberfire</groupId>
+      <artifactId>uberfire-rest-client</artifactId>
     </dependency>
   </dependencies>
 </project>
