+++ test-cloud/test-cloud-common/src/main/java/org/kie/cloud/tests/common/time/TimeUtils.java
@@ -13,7 +13,7 @@
  * limitations under the License.
 */
 
-package org.kie.cloud.integrationtests.util;
+package org.kie.cloud.tests.common.time;
 
 import java.time.Duration;
 import java.time.Instant;
