+++ test-cloud/test-cloud-remote/src/test/java/org/kie/cloud/integrationtests/architecture/EmptyManagedKieServersWithSmartRouterAndControllerIntegrationTest.java
@@ -40,12 +40,12 @@ import org.kie.cloud.common.provider.KieServerClientProvider;
 import org.kie.cloud.common.provider.KieServerControllerClientProvider;
 import org.kie.cloud.common.provider.SmartRouterAdminClientProvider;
 import org.kie.cloud.integrationtests.category.ApbNotSupported;
-import org.kie.cloud.integrationtests.util.Constants;
-import org.kie.cloud.integrationtests.util.SmartRouterUtils;
-import org.kie.cloud.integrationtests.util.TimeUtils;
-import org.kie.cloud.integrationtests.util.WorkbenchUtils;
 import org.kie.cloud.maven.MavenDeployer;
 import org.kie.cloud.maven.constants.MavenConstants;
+import org.kie.cloud.tests.common.client.util.SmartRouterUtils;
+import org.kie.cloud.tests.common.client.util.WorkbenchUtils;
+import org.kie.cloud.tests.common.time.Constants;
+import org.kie.cloud.tests.common.time.TimeUtils;
 import org.kie.server.api.model.KieContainerStatus;
 import org.kie.server.api.model.KieServerInfo;
 import org.kie.server.api.model.ReleaseId;
