+++ test-cloud/test-cloud-remote/src/test/java/org/kie/cloud/integrationtests/architecture/EmptyUnmanagedKieServersWithSmartRouterIntegrationTest.java
@@ -39,10 +39,10 @@ import org.kie.cloud.common.provider.SmartRouterAdminClientProvider;
 import org.kie.cloud.integrationtests.category.ApbNotSupported;
 import org.kie.cloud.integrationtests.category.JBPMOnly;
 import org.kie.cloud.integrationtests.category.OperatorNotSupported;
-import org.kie.cloud.integrationtests.util.Constants;
-import org.kie.cloud.integrationtests.util.SmartRouterUtils;
-import org.kie.cloud.integrationtests.util.TimeUtils;
 import org.kie.cloud.maven.constants.MavenConstants;
+import org.kie.cloud.tests.common.client.util.SmartRouterUtils;
+import org.kie.cloud.tests.common.time.Constants;
+import org.kie.cloud.tests.common.time.TimeUtils;
 import org.kie.server.api.model.KieContainerResource;
 import org.kie.server.api.model.KieContainerStatus;
 import org.kie.server.api.model.ReleaseId;
