+++ test-cloud/test-cloud-remote/src/test/java/org/kie/cloud/integrationtests/testproviders/HttpsKieServerTestProvider.java
@@ -30,8 +30,8 @@ import cz.xtf.client.HttpResponseParser;
 import org.apache.http.entity.ContentType;
 import org.assertj.core.api.Assertions;
 import org.kie.cloud.api.deployment.KieServerDeployment;
-import org.kie.cloud.integrationtests.Kjar;
 import org.kie.cloud.maven.MavenDeployer;
+import org.kie.cloud.tests.common.client.util.Kjar;
 import org.kie.server.api.KieServerConstants;
 import org.kie.server.api.marshalling.Marshaller;
 import org.kie.server.api.marshalling.MarshallerFactory;
