+++ test-cloud/test-cloud-remote/src/test/java/org/kie/cloud/integrationtests/testproviders/OptaplannerTestProvider.java
@@ -25,9 +25,9 @@ import org.kie.api.KieServices;
 import org.kie.api.runtime.KieContainer;
 import org.kie.cloud.api.deployment.KieServerDeployment;
 import org.kie.cloud.common.provider.KieServerClientProvider;
-import org.kie.cloud.integrationtests.Kjar;
-import org.kie.cloud.integrationtests.util.KieServerUtils;
 import org.kie.cloud.maven.MavenDeployer;
+import org.kie.cloud.tests.common.client.util.KieServerUtils;
+import org.kie.cloud.tests.common.client.util.Kjar;
 import org.kie.server.api.model.KieContainerResource;
 import org.kie.server.api.model.ReleaseId;
 import org.kie.server.api.model.instance.SolverInstance;
